#13 Verificando se usuário está logado.

+++ app/Controller/ContasController.php
@@ -1,51 +1,75 @@
 <?php
 require_once '../Model/Contas.php';
+require_once 'AuthController.php';
 
 class ContasController
 {
     private $contas;
+    private $auth;
 
     function __construct()
     {
         $this->contas = new Contas();
+        $this->auth = new AuthController();
     }
 
     public function index()
     {
-        return $this->contas->allContas();
+        if($this->auth->verify_logged()) {
+            return $this->contas->allContas();
+        } else{
+            return "Favor realizar login.";
+        }
     }
 
     public function show($id)
     {
-        return $this->contas->getConta($id);
+        if($this->auth->verify_logged()) {
+            return $this->contas->getConta($id);
+        } else{
+            return "Favor realizar login.";
+        }
     }
 
     public function store()
     {
-        $conta = "Gerente";
-        $sigla_conta = "ger";
-        $data = array(
-            "desc_conta" => $conta,
-            "sigla_conta" => $sigla_conta
-        );
-        return $this->contas->addConta($data);
+        if($this->auth->verify_logged()) {
+            $conta = "Gerente";
+            $sigla_conta = "ger";
+            $data = array(
+                "desc_conta" => $conta,
+                "sigla_conta" => $sigla_conta
+            );
+            return $this->contas->addConta($data);
+        } else{
+            return "Favor realizar login.";
+        }
     }
 
     public function update($id)
     {
-        $conta = "Supervisor";
-        $sigla_conta = "sup";
-        $data = array(
-            "desc_conta" => $conta,
-            "sigla_conta" => $sigla_conta,
-            "conta_id" => $id,
-        );
-        return $this->contas->updateConta($data);
+        if($this->auth->verify_logged()) {
+            $conta = "Supervisor";
+            $sigla_conta = "sup";
+            $data = array(
+                "desc_conta" => $conta,
+                "sigla_conta" => $sigla_conta,
+                "conta_id" => $id,
+            );
+            return $this->contas->updateConta($data);
+        } else{
+            return "Favor realizar login.";
+        }
     }
 
     public function destroy($id)
     {
-        return $this->contas->deleteConta($id);
+        if($this->auth->verify_logged()) {
+            return $this->contas->deleteConta($id);
+        } else{
+            return "Favor realizar login.";
+        }
     }
-
 }
+$obj = new ContasController();
+var_dump($obj->destroy(1));
